aoc2025 explicit integer types

+++ advent_of_code/2025/solutions/src/day_08.rs
@@ -10,13 +10,13 @@ pub struct Data(Vec<Point>);
 
 #[derive(Debug, Clone)]
 struct Point {
-    x: usize,
-    y: usize,
-    z: usize,
+    x: u32,
+    y: u32,
+    z: u32,
 }
 
 impl Point {
-    fn dist_squared(&self, other: &Point) -> usize {
+    fn dist_squared(&self, other: &Point) -> u32 {
         self.x.abs_diff(other.x).pow(2)
             + self.y.abs_diff(other.y).pow(2)
             + self.z.abs_diff(other.z).pow(2)
